Make cpio files reproducible
When packing directories into cpio files we want those to be bit
identical from one run to another.

This can be done by using HEAD commit time for all files, but this would
result in a changed cpio file if tracking a branch and subdirectory
didn't change.

To avoid that we want each file to have it's mtime set to the time of
the last commit that touched that file.

A side effect of this is that we need to do a deep clone to have access
to the git history.

Signed-off-by: Nicolas Belouin <nicolas.belouin@suse.com>