Use the tree hash for the service info and always write the obsinfo file
Querying the tree hash is much faster than generating an md5sum for
a directory tree. We could not do this before as the clone method
threw away the git data if a subdir was used.

Also, this commit changes the code to also write an obsinfo file
if a subdir is used. In this case, we query git for the last
commit that changed the subdir. Unfortunately we can no longer
do a shallow clone, as we need the trees of the last commits.